requirements.txt
langgraph
langchain
langchain-core
langchain-openai
langchain-community
qdrant-client
openpyxl
pydantic
python-dotenv
pandas
pyyaml
pypdf
python-docx
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib